Polish feedback & checkpoint design
